Remove redundant code; add logs
The check `if result.ok` is not needed since `raise_for_status` would
have prevented ever getting to a result that is not OK.

Also, in order to add debug logging, the client request is wrapped in
the try/except block so that a debug log can be printed. The original
exception is just re-raised - so no functionality change other than the
debug log.