Set non-numeric hourly forecast wind speeds to 0
This is to help address issue #83 . The XML files that are generated by the service can sometimes have a wind speed value of "Calm" or "Calme" depending on the language.

After looking at many of the available files, the only wind speed values I saw used were 5, 10, 15, 20, 30, 40, and "Calm" or "Calme". I'm assuming the "Calm" value is effectively "0".