Cleanup precip_type logic.
Now the caching code is more integrated in this module we can cache both
snow/rain types at the same time.

Also, renamed the precip variable to reflect closer to how they are
used.

Finally, precip_type getter is a breaking change. HA does not use it so
does no change required for it.