Update exception handling
Now more consistent in that the code will always raise
ECWeatherUpdateFailed exception.

This will allow users of the API to always catch that one exception.

Prior, raise_for_status would raise an aiohttp error.

In addition to the consistency, the text in the exception has been
tidied a bit and more info given.